refactor(parser): split stmt/simple.rs into cohesive submodules (§7-8)
Split the 2167-line parser/stmt/simple.rs (over the 500-line limit) into a
facade plus eight themed submodules under src/parser/stmt/simple/, all
under 500 lines. simple.rs is free parser functions addressed as
simple::<name> (~205 call sites); the facade keeps all thread_local!/static
parser-state registries and re-exports each submodule preserving its
original visibility, so no external caller changes. stmt/mod.rs is untouched.

- simple.rs (288)           - facade: use lines, parser-state statics, re-exports
- control_stmts.rs (425)    - return/last/next/redo/goto/die/take/catch/phaser/block/known_call
- module_exports.rs (390)   - module/inline-module export registration + import
- user_ops.rs (386)         - match_user_declared_* operator/term lookups
- io_stmts.rs (218)         - say/print/put/note + parse_expr_list
- registry.rs (187)         - user-sub / op-precedence registry accessors
- lib_paths.rs (158)        - parser lib/program path setters
- compile_consts.rs (96)    - compile-time constants + scope push/pop
- pragma_preseed.rs (70)    - attributes pragma + eval preseeds

Pure relocation. Descendant submodules see the facade's private statics, so
no static visibility bumps were needed. Moved fns are pub(crate) in
submodules (true pub for the three crate-exported lib-path setters) and
re-exported at their original visibility; private helpers moved with their
callers. Body refs to stmt-level siblings gained one super:: level. The fn
multiset is byte-identical. No behavior change.
